Fix (gallery): Fix arrow key navigation to scroll at most one row in tile mode
Previously, in tile mode, arrow key navigation always scroll the page such that the active image is at the top.

Now, in tile mode, arrow key navigation only scrolls the page when the active image is outside the scope of current view, and scrolls only the height of one row.